tests/jax/agentic_behavior/test_agentic_behavior.py
import pytest
import jax.numpy as jnp
from RL_Developments.Jax.agentic_behavior import (
    NeuroFlexAgenticBehavior,
    ZeroShotAgent,
    FewShotAgent,
    ChainOfThoughtAgent,
    MetaPromptingAgent,
    SelfConsistencyAgent,
    GenerateKnowledgePromptingAgent,
    create_neuroflex_agentic_behavior
)
import flax.linen as nn

class MockModel(nn.Module):
    def apply(self, params, x):
        return x

def mock_tokenizer(text):
    return [ord(c) for c in text]

@pytest.fixture
def neuroflex_agent():
    model = MockModel()
    tokenizer = mock_tokenizer
    vocab_size = 256
    return create_neuroflex_agentic_behavior(model, tokenizer, vocab_size)

def test_zero_shot(neuroflex_agent):
    prompt = "What is the capital of France?"
    response = neuroflex_agent.zero_shot(prompt)
    assert isinstance(response, str)

def test_few_shot(neuroflex_agent):
    prompt = "Translate 'hello' to French."
    examples = [{"input": "Translate 'cat' to French.", "output": "chat"}]
    response = neuroflex_agent.few_shot(prompt, examples)
    assert isinstance(response, str)

def test_chain_of_thought(neuroflex_agent):
    prompt = "Solve the equation: 2 + 2"
    response = neuroflex_agent.chain_of_thought(prompt)
    assert isinstance(response, str)

def test_meta_prompting(neuroflex_agent):
    prompt = "Write a poem about the sea."
    meta_prompt = "Use a rhyming scheme."
    response = neuroflex_agent.meta_prompting(prompt, meta_prompt)
    assert isinstance(response, str)

def test_self_correct(neuroflex_agent):
    output = "The capital of France is Berlin."
    corrected_output = neuroflex_agent.self_correct(output)
    assert isinstance(corrected_output, str)

def test_self_update(neuroflex_agent):
    feedback = "The capital of France is Paris, not Berlin."
    neuroflex_agent.self_update(feedback)
    assert True  # Ensure no exceptions are raised

def test_plan_and_execute(neuroflex_agent):
    task = "Plan a trip to Paris."
    plan, result = neuroflex_agent.plan_and_execute(task)
    assert isinstance(plan, list)
    assert isinstance(result, str)

def test_multi_agent_collaboration(neuroflex_agent):
    task = "Write a story about a hero."
    other_agents = [neuroflex_agent, neuroflex_agent]
    response = neuroflex_agent.multi_agent_collaboration(task, other_agents)
    assert isinstance(response, str)

if __name__ == "__main__":
    pytest.main()
